backend/requirements.txt
fastapi>=0.68.0,<0.69.0
uvicorn>=0.15.0,<0.16.0
transformers==4.33.2
torch==2.0.1
python-Levenshtein==0.21.1
nltk==3.8.1